% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{biodtshiny_example}
\alias{biodtshiny_example}
\title{Get path to raw data}
\usage{
biodtshiny_example(path = NULL)
}
\arguments{
\item{path}{library path}
}
\description{
Get path to raw data
}
