% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_beehave_fct.R
\name{init_input_parameters_beehave}
\alias{init_input_parameters_beehave}
\title{Title}
\usage{
init_input_parameters_beehave()
}
\value{
list, list containing used constant values
}
\description{
Title
}
