% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auth.R
\name{mod_login_server}
\alias{mod_login_server}
\title{login server module}
\usage{
mod_login_server(id, r)
}
\arguments{
\item{id}{session id}
}
\value{
The module will return a reactive 2 element list to your main application.
  First element \code{user_auth} is a boolean indicating whether there has been
  a successful login or not. Second element \code{info} will be the data frame provided
  to the function, filtered to the row matching the successfully logged in username.
  When \code{user_auth} is FALSE \code{info} is NULL.
}
\description{
Shiny authentication module for use with \link{mod_login_ui}
}
\details{
Call via \code{shiny::callModule("your_id", ...)}
}
\author{
Tomas Martinovic \email{tomas.martinovic@vsb.cz}
}
