% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auth.R
\name{mod_login_ui}
\alias{mod_login_ui}
\title{login UI module}
\usage{
mod_login_ui(
  id,
  title = "Login please",
  user_title = "Username",
  pass_title = "Password",
  login_title = "Login",
  additional_ui = NULL
)
}
\arguments{
\item{id}{Shiny id}

\item{title}{header title for the login panel}

\item{user_title}{label for the user name text input}

\item{pass_title}{label for the password text input}

\item{login_title}{label for the login button}

\item{additional_ui}{additional shiny UI element to add below login button. Wrap multiple inside \code{shiny::tagList()}}

\item{error_message}{message to display after failed login}
}
\value{
Shiny UI
}
\description{
Shiny UI Module for use with \link{mod_login_server}
This is taken from the shinyauthr package authored by Paul Campbell, \email{pacampbell91@gmail.com} and found at \url{https://github.com/PaulC91/shinyauthr}.
Mainly the login logic was adjusted for the needs of this project.
}
\details{
Call via \code{mod_login_ui("your_id")}
}
\author{
Tomas Martinovic \email{tomas.martinovic@vsb.cz}
}
